make `separate module` for book class
